Bot_App/requirements.txt
logging
mysql-connector-python 